hotfix: separate logic from Handler
