+++ src/main/java/no/unit/nva/doi/transformer/MainHandler.java
@@ -1,6 +1,5 @@
 package no.unit.nva.doi.transformer;
 
-import static no.unit.nva.model.util.OrgNumberMapper.UNIT_ORG_NUMBER;
 import static org.apache.http.HttpHeaders.CONTENT_LOCATION;
 import static org.apache.http.HttpHeaders.CONTENT_TYPE;
 import static org.apache.http.HttpStatus.SC_BAD_REQUEST;
@@ -12,7 +11,6 @@ import static org.zalando.problem.Status.INTERNAL_SERVER_ERROR;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestStreamHandler;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -23,34 +21,20 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
-import java.net.URI;
-import java.time.Instant;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
-import no.unit.nva.doi.transformer.model.crossrefmodel.CrossRefDocument;
-import no.unit.nva.doi.transformer.model.crossrefmodel.CrossrefApiResponse;
-import no.unit.nva.doi.transformer.model.internal.external.DataciteResponse;
 import no.unit.nva.model.Publication;
-import no.unit.nva.model.util.OrgNumberMapper;
 import org.apache.http.entity.ContentType;
 import org.zalando.problem.Problem;
 import org.zalando.problem.ProblemModule;
 
 public class MainHandler implements RequestStreamHandler {
 
-    public static final String ORG_NUMBER_COUNTRY_PREFIX_NORWAY = "NO";
     public static final String ACCESS_CONTROL_ALLOW_ORIGIN = "Access-Control-Allow-Origin";
     public static final String BODY = "body";
     public static final String HEADERS = "headers";
-
-    public static final String REQUEST_CONTEXT_AUTHORIZER_CLAIMS = "/requestContext/authorizer/claims/";
-    public static final String CUSTOM_FEIDE_ID = "custom:feideId";
-    public static final String CUSTOM_ORG_NUMBER = "custom:orgNumber";
-    public static final String MISSING_CLAIM_IN_REQUEST_CONTEXT = "Missing claim in requestContext: ";
 
     public static final String ALLOWED_ORIGIN = "ALLOWED_ORIGIN";
     public static final String APPLICATION_PROBLEM = "application/problem+json";
@@ -85,19 +69,14 @@ public class MainHandler implements RequestStreamHandler {
 
     @Override
     public void handleRequest(InputStream input, OutputStream output, Context context) throws IOException {
-        String owner;
-        String orgNumber;
+        JsonNode event;
         String body;
         String contentLocation;
         try {
-            String result = new BufferedReader(new InputStreamReader(input))
-                .lines().collect(Collectors.joining("\n"));
-            JsonNode event = objectMapper.readTree(result);
-
+            String result = new BufferedReader(new InputStreamReader(input)).lines().collect(Collectors.joining("\n"));
+            event = objectMapper.readTree(result);
             body = extractRequestBody(event);
             contentLocation = extractContentLocationHeader(event);
-            owner = getClaimValueFromRequestContext(event, CUSTOM_FEIDE_ID);
-            orgNumber = getClaimValueFromRequestContext(event, CUSTOM_ORG_NUMBER);
         } catch (Exception e) {
             e.printStackTrace();
             objectMapper.writeValue(output,
@@ -107,11 +86,9 @@ public class MainHandler implements RequestStreamHandler {
         }
 
         try {
-            UUID uuid = UUID.randomUUID();
-            URI publisherID = toPublisherId(orgNumber);
-            Instant now = Instant.now();
-            Publication publication = convertInputToPublication(body, contentLocation, now, owner, uuid, publisherID);
-
+            PublicationTransformer transformer = new PublicationTransformer(dataciteConverter, crossRefConverter,
+                createObjectMapper());
+            Publication publication = transformer.transformPublication(event, body, contentLocation);
             log(objectMapper.writeValueAsString(publication));
             objectMapper.writeValue(output,
                 new GatewayResponse<>(objectMapper.writeValueAsString(publication), sucessResponseHeaders(), SC_OK));
@@ -130,30 +107,6 @@ public class MainHandler implements RequestStreamHandler {
         } else {
             return body.textValue();
         }
-    }
-
-    protected Publication convertInputToPublication(String body, String contentLocation, Instant now, String owner,
-                                                    UUID identifier, URI publisher) throws JsonProcessingException {
-
-        MetadataLocation metadataLocation = MetadataLocation.lookup(contentLocation);
-        if (metadataLocation.equals(MetadataLocation.CROSSREF)) {
-            return convertFromCrossRef(body, now, owner, identifier, publisher);
-        } else {
-            return convertFromDatacite(body, now, owner, identifier, publisher);
-        }
-    }
-
-    private Publication convertFromDatacite(String body, Instant now, String owner, UUID uuid, URI publisherId)
-        throws JsonProcessingException {
-        DataciteResponse dataciteResponse = objectMapper.readValue(body, DataciteResponse.class);
-        return dataciteConverter.toPublication(dataciteResponse, now, uuid, owner, publisherId);
-    }
-
-    private Publication convertFromCrossRef(String body, Instant now, String owner, UUID identifier, URI publisherId)
-        throws JsonProcessingException {
-
-        CrossRefDocument document = objectMapper.readValue(body, CrossrefApiResponse.class).getMessage();
-        return crossRefConverter.toPublication(document, now, owner, identifier, publisherId);
     }
 
     private String extractContentLocationHeader(JsonNode event) {
@@ -203,18 +156,5 @@ public class MainHandler implements RequestStreamHandler {
 
     public static void log(String message) {
         System.out.println(message);
-    }
-
-    private URI toPublisherId(String orgNumber) {
-        if (orgNumber.startsWith(ORG_NUMBER_COUNTRY_PREFIX_NORWAY)) {
-            // Remove this if and when datamodel has support for OrgNumber country prefix
-            return OrgNumberMapper.toCristinId(orgNumber.substring(ORG_NUMBER_COUNTRY_PREFIX_NORWAY.length()));
-        }
-        return OrgNumberMapper.toCristinId(orgNumber);
-    }
-
-    private String getClaimValueFromRequestContext(JsonNode event, String claimName) {
-        return Optional.ofNullable(event.at(REQUEST_CONTEXT_AUTHORIZER_CLAIMS + claimName).textValue())
-                       .orElse(UNIT_ORG_NUMBER);
     }
 }
